@@ -1,14 +1,14 @@
 name: ical_parser
 description: A parser for ICalendar files (.ics)
-version: 1.0.1
+version: 1.1.0
 repository: https://github.com/julisanchez/ical-parser
 
 environment:
-  sdk: '>=2.8.1 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 #dependencies:
 #  path: ^1.7.0
 
 dev_dependencies:
-  pedantic: ^1.9.0
-  test: ^1.14.4
+  pedantic: ^1.11.0
+  test: ^1.17.3
